Allow setting folderUid
When running a deployment with more that one replica, the generated
dashboard folder UIDs can be different per instance. This creates
inconsistent behavior when using the "Go to folder" links in
`/dashboards/` as, depending on which instance the request hits, the
user might get a "Folder not found" error.

The new `folderUidGenerator` function can be overriden to generate
consistent folder UIDs by setting `folderUid` in the dashboard provider
(docs [here][docs], example below). The default behavior of not setting
a UID was maintained to not break expectations.

Additionally, I noticed while testing this feature that Grafana only
loads the dashboard providers on startup, so I added a checksum for the
providers configmap.

```jsonnet
local grafana = import 'grafana/grafana.libsonnet';

local grafanaWithDashboards =
  (grafana
   {
     _config+:: {
       namespace: 'monitoring-grafana',
       grafana+:: {
         folderDashboards+:: (import 'dashboard_folders.libsonnet'),
         folderUidGenerator(folder):: std.md5(folder),
       },
     },
   }).grafana;

{
  apiVersion: 'v1',
  kind: 'List',
  items:
    grafanaWithDashboards.dashboardDefinitions +
    [
      grafanaWithDashboards.dashboardSources,
      grafanaWithDashboards.dashboardDatasources,
      grafanaWithDashboards.deployment {
        spec+: { replicas: 3 },
      },
      grafanaWithDashboards.serviceAccount,
      grafanaWithDashboards.service,
    ],
}
```

[docs]: https://grafana.com/docs/grafana/latest/administration/provisioning/#dashboards